pyproject.toml
[project]
name = "form2list"
version = "0.1.2"
description = "A script to process Excel files and generate lists."
authors = [
    {"name"="Mitsuru Nakakawaji"},
    {"email"="mitsuru@procube.jp"}
]
license = {file="LICENSE"}
readme = "README.md"

dependencies = [
    "openpyxl",
    "pyyaml",
    "jinja2"
]

[project.scripts]
form2list = "form2list:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest"
]
